+++ .gitignore
@@ -1,1 +1,7 @@
 *.user
+
+# AppImages
+*.AppImage
+
+# build results
+*build*/
