+++ resources/CMakeLists.txt
@@ -1,0 +1,8 @@
+install(
+    FILES asandbox.desktop
+    DESTINATION share/applications
+)
+install(
+    FILES asandbox.svg
+    DESTINATION share/icons/hicolor/scalable/apps
+)
